requirements.txt
sqlalchemy==2.0.23
pandas==2.0.3
openpyxl==3.1.2
fastapi==0.104.1
jinja2==3.1.2
pyodbc==5.0.1
python-dotenv==1.0.0
uvicorn==0.24.0
aiofiles==23.2.1
python-multipart==0.0.6
pandastable==0.13.1
ttkthemes==3.2.2
matplotlib==3.7.1
pillow>=9.0.0